requirements.txt
amqp==5.1.1
aniso8601==9.0.1
billiard==4.1.0
blinker==1.6.2
celery==5.3.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.3.0
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
prompt-toolkit==3.0.39
PyJWT==2.7.0
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.9
Werkzeug==2.3.3
